refactor: Include `sink_frontend` in `log_fabric`
The `sink_frontend.hpp` header has become sufficiently lightweight and
overlapping in terms of includes that including it in `log_fabric.hpp`
is viable again. The reduction in header heft has been achieved by
splitting the `file_sink` into its own header.
